requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
httpx
alembic 
psycopg2-binary 
asyncpg 
python-dotenv